new integration test - pre
