+++ e2e-tests/src/test/kotlin/ru/racelog3/e2e_tests/AuthBaseTest.kt
@@ -26,5 +26,7 @@ abstract class AuthBaseTest : BaseTest() {
 
         checkText(createID(RACER_PROFILE_NAME, "1"), "", "Check the name of new racer profile")
         checkText(createID(RACER_PROFILE_BIKE_NUMBER, "1"), "", "Check the bike number of new racer profile")
+
+        clickElement(ADMIN_RACE_INFO_EXPAND_BUTTON)
     }
 }
